requirements.txt
asgiref==3.3.4
Django==3.2.3
pytz==2021.1
sqlparse==0.4.1

# pre-commit hooks
pre-commit

# To load env variables
python-dotenv==0.17.1

# for different env configurations
django-configurations==2.2

# For websockets
channels==3.0.3
uvicorn==0.13.4

# Rest APIs
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1

#CORS
django-cors-headers==3.7.0

# Developer tools
ipython==7.24.0 # interactive python console
ipdb==0.13.8 # interactive python debugger
flake8==3.9.2 # Code formatter
mkdocs==1.1.2

# Database related
psycopg2==2.8.6 # postgre Adaptor for Django
pillow==8.2.0 # if we want to store images in DB


# To validate json field schema
jsonschema==3.2.0
